Code review: Add test for notifications
